+++ saml/plugins/auth/saml/saml.auth.php
@@ -1,7 +1,7 @@
 <?php
 /**
 * @author     Laurent Jouanneau
-* @copyright  2019-2022 3liz
+* @copyright  2019-2026 3liz
 * @licence  http://www.gnu.org/licenses/lgpl.html GNU Lesser General Public Licence, see LICENCE file
 */
 
@@ -17,6 +17,8 @@ class samlAuthDriver extends jAuthDriverBase implements jIAuthDriver3 {
      * @var bool true if the current login process is made with SAML
      */
     protected $authWithSamlActivated = false;
+
+    protected $loginsCache = array();
 
     function __construct($params) {
         if (property_exists(jApp::config(), 'saml') && is_array(jApp::config()->saml)) {
@@ -40,6 +42,10 @@ class samlAuthDriver extends jAuthDriverBase implements jIAuthDriver3 {
         return jDao::get($this->_params['dao'], $this->_params['profile']);
     }
 
+    /**
+     * @param jDaoRecordBase $user
+     * @return true
+     */
     public function saveNewUser($user)
     {
         $this->getDao()->insert($user);
@@ -51,6 +57,9 @@ class samlAuthDriver extends jAuthDriverBase implements jIAuthDriver3 {
         return true;
     }
 
+    /**
+     * @param jDaoRecordBase $user
+     */
     protected function createSAMLAccount($user)
     {
         $samlAccount = jDao::createRecord('saml~saml_account');
@@ -65,6 +74,12 @@ class samlAuthDriver extends jAuthDriverBase implements jIAuthDriver3 {
     public function removeUser($login)
     {
         $dao = $this->getDao();
+        // retrieve the real login, in the case where the given login does not match character case
+        $user = $this->getUser($login);
+        if (!$user) {
+            return false;
+        }
+
         if (function_exists('mb_strtolower')) {
             $loginLegacy = mb_strtolower($login);
             if ($loginLegacy != $login) {
@@ -75,12 +90,15 @@ class samlAuthDriver extends jAuthDriverBase implements jIAuthDriver3 {
             }
         }
 
-        jDao::get('saml~saml_account')->delete($login);
-
-        $dao->deleteByLogin($login);
+        jDao::get('saml~saml_account')->delete($user->login);
+        $dao->deleteByLogin($user->login);
+        unset($this->loginsCache[$login]);
         return true;
     }
 
+    /**
+     * @param jDaoRecordBase $user
+     */
     public function updateUser($user)
     {
         if (!is_object($user)) {
@@ -95,22 +113,47 @@ class samlAuthDriver extends jAuthDriverBase implements jIAuthDriver3 {
         return true;
     }
 
+    /**
+     * Retrieve the jAuth account with case-insensitive matching
+     *
+     * @param $login
+     * @return jDaoRecordBase|null
+     */
     public function getUser($login)
     {
-        $dao = $this->getDao();
-        $user = $dao->getByLogin($login);
-        if ($user) {
-            return $user;
+        if (isset($this->loginsCache[$login])) {
+            return $this->loginsCache[$login];
         }
+
         if (function_exists('mb_strtolower')) {
-            // compatibility with an old version of the module, when
-            // logins were stored only in lowercase
-            $loginLegacy = mb_strtolower($login);
-            if ($loginLegacy != $login) {
-                return $dao->getByLogin($loginLegacy);
+            $llogin = mb_strtolower($login);
+        }
+        else {
+            $llogin = strtolower($login);
+        }
+        $conditions = new jDaoConditions();
+        $conditions->addCondition('login', '=', $llogin, 'lower(%s)');
+        $rs = $this->getDao()->findBy($conditions);
+
+        // IDP may return a login that have a different character case than the login of a previous
+        // authentication, so previous versions of the module may create several accounts with the
+        // same login having different character case.
+        // So, if there are multiple account with the same login having different character case,
+        // we return the one which match exactly the given login, else we return another one.
+        $alternateUser = null;
+        foreach ($rs as $rec) {
+            if ($rec->login == $login) {
+                $this->loginsCache[$login] = $rec;
+                return $rec;
+            }
+            else if ($alternateUser == null) {
+                $alternateUser = $rec;
             }
         }
-        return false;
+        if ($alternateUser) {
+            $this->loginsCache[$login] = $alternateUser;
+        }
+        return $alternateUser;
     }
 
     public function createUserObject($login, $password)
@@ -121,7 +164,8 @@ class samlAuthDriver extends jAuthDriverBase implements jIAuthDriver3 {
         return $user;
     }
 
-    public function getUserList($pattern) {
+    public function getUserList($pattern)
+    {
         $dao = $this->getDao();
         if ($pattern == '%' || $pattern == '') {
             return $dao->findAll();
@@ -134,7 +178,12 @@ class samlAuthDriver extends jAuthDriverBase implements jIAuthDriver3 {
 
     public function canChangePassword($login)
     {
-        $can = $this->canUseLocalPassword($login);
+        $user = $this->getUser($login);
+        $can = false;
+        if ($user) {
+            $can = $this->canUseLocalPassword($user->login);
+        }
+
         if (!$can) {
             $this->reasonToForbiddenPasswordChange = jLocale::get('saml~auth.localpassword.error');
         }
@@ -152,7 +201,11 @@ class samlAuthDriver extends jAuthDriverBase implements jIAuthDriver3 {
 
     public function changePassword($login, $newpassword)
     {
-        return $this->getDao()->updatePassword($login, $this->cryptPassword($newpassword));
+        $user = $this->getUser($login);
+        if ($user) {
+            return $this->getDao()->updatePassword($user->login, $this->cryptPassword($newpassword));
+        }
+        return false;
     }
 
     /**
@@ -186,7 +239,7 @@ class samlAuthDriver extends jAuthDriverBase implements jIAuthDriver3 {
     public function verifyPassword($login, $password)
     {
         $daouser = $this->getDao();
-        $user = $daouser->getByLogin($login);
+        $user = $this->getUser($login);
 
         if ($this->authWithSamlActivated) {
             // authentication with SAML
@@ -211,7 +264,7 @@ class samlAuthDriver extends jAuthDriverBase implements jIAuthDriver3 {
             }
             else {
                 // update saml account data
-                $samlAccount = jDao::get('saml~saml_account')->get($login);
+                $samlAccount = jDao::get('saml~saml_account')->get($user->login);
                 if ($samlAccount) {
                     $samlAccount->lastUsed = date('Y-m-d H:i:s');
                     $samlAccount->usageCount++;
@@ -229,7 +282,7 @@ class samlAuthDriver extends jAuthDriverBase implements jIAuthDriver3 {
                 return false;
             }
 
-            if (!$this->canUseLocalPassword($login)) {
+            if (!$this->canUseLocalPassword($user->login)) {
                 return false;
             }
 
@@ -240,7 +293,7 @@ class samlAuthDriver extends jAuthDriverBase implements jIAuthDriver3 {
             if ($result !== true) {
                 // it is a new hash for the password, let's update it persistently
                 $user->password = $result;
-                $daouser->updatePassword($login, $result);
+                $daouser->updatePassword($user->login, $result);
             }
         }
         return $user;
@@ -274,14 +327,17 @@ class samlAuthDriver extends jAuthDriverBase implements jIAuthDriver3 {
     public function getAuthenticationPermissions($login)
     {
         $perms = self::AUTH_NOT_ALLOWED;
-
-        $samlAccount = jDao::get('saml~saml_account')->get($login);
+        $user = $this->getUser($login);
+        if (!$user) {
+            return $perms;
+        }
+        $samlAccount = jDao::get('saml~saml_account')->get($user->login);
         if ($samlAccount) {
             $perms |= self::AUTH_SAML_ALLOWED;
         }
         else {
             // if the user has the right to administrate SAML configuration, it can login with local password
-            if (jAcl2::checkByUser($login, 'saml.config.access')) {
+            if (jAcl2::checkByUser($user->login, 'saml.config.access')) {
                 $perms |= self::AUTH_PASSWORD_ALLOWED;
             }
         }